# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Zia Carbon Footprint API",
        "description": "API documentation for the Zia Carbon Footprint Dashboard and Backend logic.",
        "contact": {
            "email": "support@zia.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Dynamic Host"
        }
    ],
    "paths": {
        "/api/admin/companies": {
            "get": {
                "tags": [
                    "Admin - Companies"
                ],
                "summary": "List all companies with periods (Admin)",
                "operationId": "aa7f49a5bd3c35517ffdaa353feb612f",
                "responses": {
                    "200": {
                        "description": "List of companies"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin - Companies"
                ],
                "summary": "Create a new company",
                "operationId": "f1c7e1b7eac446bd837cfa44833139a5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "nit"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Acme Corp"
                                    },
                                    "nit": {
                                        "type": "string",
                                        "example": "123456789-0"
                                    },
                                    "sector_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "logo_url": {
                                        "type": "string",
                                        "example": "https://example.com/logo.png"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/companies/{company}": {
            "put": {
                "tags": [
                    "Admin - Companies"
                ],
                "summary": "Update a company",
                "operationId": "e8afe8dcf21bf1c0cc08846aae1ef28b",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin - Companies"
                ],
                "summary": "Soft delete a company",
                "operationId": "6ddd8e0a4568489ad8faf1b640e5bbae",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/companies/{company}/periods": {
            "post": {
                "tags": [
                    "Admin - Companies"
                ],
                "summary": "Add a new period to a company",
                "operationId": "d07bb96c70a64839dd389f2ac5f4dc5b",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "year",
                                    "status"
                                ],
                                "properties": {
                                    "year": {
                                        "type": "integer",
                                        "example": 2024
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "open"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Period added"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/periods/{period}": {
            "put": {
                "tags": [
                    "Admin - Companies"
                ],
                "summary": "Update a period",
                "operationId": "b13505062ae7d9fe6fb1b63702e37093",
                "parameters": [
                    {
                        "name": "period",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin - Companies"
                ],
                "summary": "Delete a period",
                "operationId": "7cce2efa3c2f9b7a083906e2142e56f1",
                "parameters": [
                    {
                        "name": "period",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/sectors": {
            "get": {
                "tags": [
                    "Admin - Sectors"
                ],
                "summary": "List all company sectors",
                "operationId": "c03e57c3fd398f2691b8036d15c00867",
                "responses": {
                    "200": {
                        "description": "List of sectors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin - Sectors"
                ],
                "summary": "Create a new company sector",
                "operationId": "c95c19c85d651801332d50e636df4bbe",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Industrial"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Industrial and manufacturing sector"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/sectors/{sector}": {
            "get": {
                "tags": [
                    "Admin - Sectors"
                ],
                "summary": "Get sector details",
                "operationId": "f35e5b0e723177e11764364882e1456f",
                "parameters": [
                    {
                        "name": "sector",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sector details"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin - Sectors"
                ],
                "summary": "Update a company sector",
                "operationId": "7819322d5054c4e1cc82639b55453827",
                "parameters": [
                    {
                        "name": "sector",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Industrial Updated"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin - Sectors"
                ],
                "summary": "Delete a company sector",
                "operationId": "840b5b3ef746c3af47a2b965c2ed5c07",
                "parameters": [
                    {
                        "name": "sector",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}